refactor: move Suggestion, Suggestion and Dictionary to separate files
